requirements.txt
absl-py==0.12.0
aenum==3.1.0
aiofiles==0.7.0
asgiref==3.3.4
astor==0.8.1
astunparse==1.6.3
boto3==1.17.87
botocore==1.20.87
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dataclasses-json==0.5.3
fastapi==0.65.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.5.0
Jinja2==3.0.1
jmespath==0.10.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
matplotlib==3.4.2
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
pyproj==3.1.0
python-dateutil==2.8.1
python-multipart==0.0.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
scipy==1.4.1
sentinelhub==3.3.0
six==1.15.0
starlette==0.14.2
stringcase==1.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2021.4.8
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.5
utm==0.7.0
uvicorn==0.14.0
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
